movido el detail de compra (#189)
